refactor(network): move around SwarmDriver code
